% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_measures.R
\name{change_measures}
\alias{change_measures}
\title{CHANGE MEASURES}
\usage{
change_measures(x, from, to)
}
\arguments{
\item{x}{a vector}

\item{from}{the current unit of x. One of: c("years","months","weeks","days","minutes","hours","kg","lbs","grams","ounces")}

\item{to}{the unit to change to. One of: c("years","months","weeks,"days","minutes","hours","kg","lbs","grams","ounces")}
}
\description{
This function identifies measurements in different units and transforms them accordingly.
Specifically, when participants are given instructions to give their response in a specific unit,
and they give another (by denoting this other unit), this function identifies these responses,
and transforms them accordingly (1-unit magnitudes only, ex. months to years or weeks, not to minutes).
}
