% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_measures.R
\name{clean_measures}
\alias{clean_measures}
\title{CLEAN MEASURES}
\usage{
clean_measures(x, type, symbols)
}
\arguments{
\item{x}{a vector}

\item{type}{one of c("minutes", "hours", "days", "weeks", "months", "years", "grams", "ounces", "kg", "lbs", "cms", "inches")}

\item{symbols}{a logical statement. If true, symbols of measurement such as '>','<','+' or '~', will be cleaned}
}
\description{
This function cleans responses of written units of measurement - specifically
units of type minutes, hours, months, years, ounces, grams, kilograms, or pounds. Participants, when given
the option of providing an alpha-numeric response, sometimes give their written response with the unit of
measurement included. This function includes the argument 'symbols' which will clean the most common notion
used by participants when approximating their responses.
}
