% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_money.R
\name{clean_money}
\alias{clean_money}
\title{Clean Money}
\usage{
clean_money(x, transform)
}
\arguments{
\item{x}{a vector}

\item{transform}{a logical statement. If TRUE, this will transcribe responses with 'k' into thousands}
}
\description{
This function cleans alpha-numeric financial responses of common notation used when participants are approximating their responses.
}
