% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_data.R
\docType{data}
\name{dl_data}
\alias{dl_data}
\title{DL_DATA}
\format{
An object of class \code{data.frame} with 15 rows and 6 columns.
}
\usage{
data(dl_data)
}
\description{
A mock data set inspired by real-life messy data!
}
\examples{
\dontrun{
data(dl_data)
}
}
\keyword{datasets}
