% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_calc.R
\name{range_calc}
\alias{range_calc}
\title{RANGE CALC}
\usage{
range_calc(x)
}
\arguments{
\item{x}{a vector}
}
\description{
This function identifies numeric responses given as a range between two numbers as denoted by a dash symbol "-",
whilst maintaining both numeric and alpha-numeric responses that are not a range.
This allows for data integrity to be maintained, so that alpha-numeric responses can be cleaned
separately depending on the nature of the response.
}
\examples{
\dontrun{
x <- c("20-24","idk,10-15","o-40","15-20","10+")
range.calc(x)
print(x)
[1] "22"   "idk,10-15"   "o-40"  "17.5"  "10+"
}
}
